+++ src/Application/Seating/RecommendationStatus.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EventFlow\Application\Seating;
+
+enum RecommendationStatus: string
+{
+    case DRAFT = 'draft';
+    case APPLIED = 'applied';
+}